stage 0 minigrid plain llm baseline reaches 0.59 with readable observation and interface disclosure
built a shared observation interpreter that turns the egocentric grid into plain text, and disclosed the goto object action interface (reach the target then done) in the prompt. this replaced the raw numeric observations that had collapsed the model into an always forward policy at 0.0 success.

full run on llama-3.3-70b over 200 episodes: success 0.59, median 5 steps, against random 0.06 and heuristic 0.0. failures split into premature done 54 and step budget timeout 28, which point at the belief and memory parts of the loop.

backbone pinned to a single openrouter model with optional max tokens, runner now executes episodes concurrently, reward logged as float, and a real failure breakdown added.

+++ src/probe/stage0/tracing.py
@@ -44,10 +44,10 @@ class TraceRow:
     structured_observation: str
     chosen_action: str
     action_source: str
-    reward: int
+    reward: float
     done: bool
     truncated: bool
-    cumulative_reward: int
+    cumulative_reward: float
     step_count: int
     success: bool
     failure_reason: str
